requirements.txt
allure-pytest==2.8.19
allure-python-commons==2.8.19
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.4
future==0.18.2
iniconfig==1.1.1
iso8601==0.1.13
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pyserial==3.4
pytest==6.1.2
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
python-can==3.3.4
python-can-remote==0.2.1
paramiko==2.7.2
